% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grassmann.R
\name{basis_gr}
\alias{basis_gr}
\title{Generate a set of ONB at a given P}
\usage{
basis_gr(P, p = NULL)
}
\arguments{
\item{P}{base}

\item{p}{rank of the projectors in the Grassmannian}
}
\description{
Generate a set of ONB at a given P
}
