% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{coord2tan}
\alias{coord2tan}
\title{Legacy alias for `tangent_from_coords(basis, coord)`}
\usage{
coord2tan(basis, coord)
}
\arguments{
\item{basis}{an (m by d by d) array of basis matrices}

\item{coord}{coordinate vector, or (n by m) matrix}
}
\description{
Legacy alias for `tangent_from_coords(basis, coord)`
}
\seealso{
[tangent_from_coords()]
}
