% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{coord_gr}
\alias{coord_gr}
\title{Legacy alias for `coords_from_basis(basis, x)`}
\usage{
coord_gr(basis, x)
}
\arguments{
\item{basis}{an (m by d by d) array of basis matrices}

\item{x}{a (d by d) matrix or (n by d by d) array}
}
\description{
Legacy alias for `coords_from_basis(basis, x)`
}
\seealso{
[coords_from_basis()]
}
