% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws_factor_model.R
\name{dyn_RFM}
\alias{dyn_RFM}
\title{Computes predictions of RFM on BWS, with factors predicted by VAR(1)}
\usage{
dyn_RFM(x, r, test_size = 1, h = 6, batch_size = NULL, max.iter = 100)
}
\description{
Computes predictions of RFM on BWS, with factors predicted by VAR(1)
}
