% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{Exp_BWS}
\alias{Exp_BWS}
\title{Legacy alias for `Exp_map(manifold_bws(), X, M)`}
\usage{
Exp_BWS(X, M)
}
\arguments{
\item{X}{tangent vector(s) at M}

\item{M}{base point}
}
\description{
Legacy alias for `Exp_map(manifold_bws(), X, M)`
}
\seealso{
[Exp_map()]
}
