% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{Exp_gr}
\alias{Exp_gr}
\title{Legacy alias for `Exp_map(manifold_grassmann(), D, P)`}
\usage{
Exp_gr(D, P)
}
\arguments{
\item{D}{tangent vector(s) at P}

\item{P}{base projector}
}
\description{
Legacy alias for `Exp_map(manifold_grassmann(), D, P)`
}
\seealso{
[Exp_map()]
}
