% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grassmann.R
\name{Exp_gr_core}
\alias{Exp_gr_core}
\title{Exponential map on Gr(d, p)}
\usage{
Exp_gr_core(D, P)
}
\arguments{
\item{D}{A projector-perspective tangent element (a symmetric matrix)}

\item{P}{A projector (base of the exponential)}
}
\description{
Exponential map on Gr(d, p)
}
