% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\name{Exp_map}
\alias{Exp_map}
\title{Exponential map on a manifold}
\usage{
Exp_map(mfd, v, mu)
}
\arguments{
\item{mfd}{a manifold object}

\item{v}{a single tangent vector or a batch of tangent vectors, at mu}

\item{mu}{base point}
}
\description{
Exponential map on a manifold
}
