% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{Exp_sphere}
\alias{Exp_sphere}
\title{Legacy alias for `Exp_map(manifold_sphere(), x, mu)`}
\usage{
Exp_sphere(x, mu)
}
\arguments{
\item{x}{tangent vector(s) at mu}

\item{mu}{base point}
}
\description{
Legacy alias for `Exp_map(manifold_sphere(), x, mu)`
}
\seealso{
[Exp_map()]
}
