% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{Exp_sphere_core}
\alias{Exp_sphere_core}
\title{Exponential map on the sphere}
\usage{
Exp_sphere_core(x, mu)
}
\arguments{
\item{x}{a tangent vector}

\item{mu}{base point}
}
\description{
Exponential map on the sphere
}
