% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws_factor_model.R
\name{Frac_Var_bws}
\alias{Frac_Var_bws}
\title{Computes evaluation metric for RFM on BWS}
\usage{
Frac_Var_bws(
  x_test,
  RFM_model,
  evaluation_type = "BWS",
  fraction = TRUE,
  Euclidean_mean,
  return_predictions = FALSE
)
}
\arguments{
\item{x_test}{raw test data}

\item{RFM_model}{output from rfm_bws}

\item{evaluation_type}{to compute BWS distance or Euclidean (Frobenius distance)}

\item{fraction}{whether to return fraction of variance unexplained or squared prediction errors}
}
\description{
Computes evaluation metric for RFM on BWS
}
