% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws_factor_model.R
\name{Frac_Var_LYB}
\alias{Frac_Var_LYB}
\title{Computes evaluation metric for linear factor model}
\usage{
Frac_Var_LYB(
  x_test,
  factor_model,
  mu_hat,
  Euclidean_mean,
  evaluation_type = "BWS",
  fraction = TRUE,
  return_predictions = FALSE,
  epsilon = 1e-06
)
}
\arguments{
\item{x_test}{raw test data}

\item{factor_model}{output from LYB_fm}

\item{mu_hat}{BWS Frechet mean, externally given}

\item{evaluation_type}{to compute BWS distance or Euclidean (Frobenius distance)}

\item{fraction}{whether to return fraction of variance unexplained or squared prediction errors}

\item{epsilon}{used in projecting predictions to SPD (only when evaluation type == "BWS")}
}
\description{
Computes evaluation metric for linear factor model
}
