% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws_factor_model.R
\name{Frac_Var_ora}
\alias{Frac_Var_ora}
\title{Computes evaluation metric for Oracle on BWS}
\usage{
Frac_Var_ora(
  x_test,
  dta,
  evaluation_type = "BWS",
  fraction = TRUE,
  Euclidean_mean,
  return_predictions = FALSE
)
}
\arguments{
\item{x_test}{raw test data}

\item{dta}{object output from dta_gen_BWS}

\item{evaluation_type}{to compute BWS distance or Euclidean (Frobenius distance)}

\item{fraction}{whether to return fraction of variance unexplained or squared prediction errors}
}
\description{
Computes evaluation metric for Oracle on BWS
}
