% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\name{geod}
\alias{geod}
\title{Geodesic distance on a manifold}
\usage{
geod(mfd, x, y)
}
\arguments{
\item{mfd}{a manifold object (manifold_bws(), manifold_grassmann(p), manifold_sphere())}

\item{x, y}{single points or (n by ...) batches of points}
}
\description{
Geodesic distance on a manifold
}
