% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{geod_BWS}
\alias{geod_BWS}
\title{Legacy alias for `geod(manifold_bws(), X, Y)`}
\usage{
geod_BWS(X, Y)
}
\arguments{
\item{X, Y}{single SPD matrices or (n by p by p) batches}
}
\description{
Legacy alias for `geod(manifold_bws(), X, Y)`
}
\seealso{
[geod()]
}
