% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws.R
\name{geod_BWS_core}
\alias{geod_BWS_core}
\title{Geodesic distance between two SPD matrices in Bures-Wasserstein}
\usage{
geod_BWS_core(X, Y)
}
\description{
Geodesic distance between two SPD matrices in Bures-Wasserstein
}
