% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grassmann.R
\name{geod_gr_core}
\alias{geod_gr_core}
\title{Geodesic distance between two projectors in Gr(d, p)}
\usage{
geod_gr_core(x, y, p = NULL)
}
\description{
Geodesic distance between two projectors in Gr(d, p)
}
