% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{geod_sphere}
\alias{geod_sphere}
\title{Legacy alias for `geod(manifold_sphere(), x, y)`}
\usage{
geod_sphere(x, y)
}
\arguments{
\item{x, y}{single points or (n by q) batches on the sphere}
}
\description{
Legacy alias for `geod(manifold_sphere(), x, y)`
}
\seealso{
[geod()]
}
