% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{Log_BWS}
\alias{Log_BWS}
\title{Legacy alias for `Log_map(manifold_bws(), X, M)`}
\usage{
Log_BWS(X, M)
}
\arguments{
\item{X}{point(s) on the manifold}

\item{M}{base point}
}
\description{
Legacy alias for `Log_map(manifold_bws(), X, M)`
}
\seealso{
[Log_map()]
}
