% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{Log_gr}
\alias{Log_gr}
\title{Legacy alias for `Log_map(manifold_grassmann(p), x, P)`}
\usage{
Log_gr(P, x, p = NULL)
}
\arguments{
\item{P}{base projector}

\item{x}{projector(s) on the manifold}

\item{p}{rank of the projectors; inferred from data if NULL}
}
\description{
Legacy alias for `Log_map(manifold_grassmann(p), x, P)`
}
\seealso{
[Log_map()]
}
