% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grassmann.R
\name{Log_gr_core}
\alias{Log_gr_core}
\title{Riemannian log map on Gr(d, p)}
\usage{
Log_gr_core(P, x, p = NULL)
}
\arguments{
\item{P}{base}

\item{x}{argument}

\item{p}{rank of the projectors in the Grassmannian}
}
\description{
Riemannian log map on Gr(d, p)
}
