% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\name{Log_map}
\alias{Log_map}
\title{Logarithm map on a manifold}
\usage{
Log_map(mfd, x, mu)
}
\arguments{
\item{mfd}{a manifold object}

\item{x}{a single point or a batch of points}

\item{mu}{base point}
}
\value{
a tangent vector, or a batch of tangent vectors, at `mu`
}
\description{
Maps a point `x` to the tangent space at base point `mu`. Inverse of
[Exp_map()].
}
\examples{
mfd = manifold_sphere()
mu = c(1, 0, 0)
x = c(0, 1, 0)
Log_map(mfd, x, mu)
}
