% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{Log_sphere}
\alias{Log_sphere}
\title{Legacy alias for `Log_map(manifold_sphere(), x, mu)`}
\usage{
Log_sphere(x, mu)
}
\arguments{
\item{x}{point(s) on the sphere}

\item{mu}{base point}
}
\description{
Legacy alias for `Log_map(manifold_sphere(), x, mu)`
}
\seealso{
[Log_map()]
}
