% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{Log_sphere_core}
\alias{Log_sphere_core}
\title{Logarithmic map for the sphere}
\usage{
Log_sphere_core(x, mu)
}
\arguments{
\item{x}{a point on the sphere}

\item{mu}{a base point}
}
\description{
Logarithmic map for the sphere
}
