% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{log_to_tangent}
\alias{log_to_tangent}
\title{Legacy alias for `tangent_from_coords(E, z)`}
\usage{
log_to_tangent(z, E)
}
\arguments{
\item{z}{coordinate vector}

\item{E}{basis array}
}
\description{
Legacy alias for `tangent_from_coords(E, z)`
}
\seealso{
[tangent_from_coords()]
}
