% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\name{map_over_points}
\alias{map_over_points}
\title{Map a single-point function over a batch of points}
\usage{
map_over_points(mfd, core_fn, x)
}
\arguments{
\item{mfd}{a manifold object}

\item{core_fn}{a function taking a single point and returning either a
point (same family as the input) or a coordinate vector}

\item{x}{a single point or a batch of points}
}
\description{
Replaces the repeated "is x a single point or an (n by ...) array of
points?" dispatch that used to live in every Exp_*/Log_* wrapper.
}
