% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{mean_on_BWS}
\alias{mean_on_BWS}
\title{Legacy alias for `frechet_mean(manifold_bws(), X, ...)`}
\usage{
mean_on_BWS(
  X,
  tau = 0.1,
  tol = 1e-04,
  max.iter = 1000,
  batch_size = NULL,
  verbose = FALSE
)
}
\arguments{
\item{X}{an (n by p by p) array of data}

\item{tau}{step size}

\item{tol}{convergence tolerance}

\item{max.iter}{maximum number of iterations}

\item{batch_size}{NULL for full-batch, or an absolute subsample count}

\item{verbose}{if TRUE, print the loss at each iteration}
}
\description{
Legacy alias for `frechet_mean(manifold_bws(), X, ...)`
}
\seealso{
[frechet_mean()]
}
