% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{mean_on_gr}
\alias{mean_on_gr}
\title{Legacy alias for `frechet_mean(manifold_grassmann(p), x, ...)`}
\usage{
mean_on_gr(
  x,
  p,
  tau = 0.1,
  tol = 1e-08,
  max.iter = 1000,
  batch_por = 1,
  init = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{an (n by d by d) array of data}

\item{p}{rank of the projectors}

\item{tau}{step size}

\item{tol}{convergence tolerance}

\item{max.iter}{maximum number of iterations}

\item{batch_por}{proportion of data subsampled per iteration (1.0 = full batch)}

\item{init}{optional initial value}

\item{verbose}{if TRUE, print the loss at each iteration}
}
\description{
Legacy alias for `frechet_mean(manifold_grassmann(p), x, ...)`
}
\seealso{
[frechet_mean()]
}
