% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{mean_on_sphere}
\alias{mean_on_sphere}
\title{Legacy alias for `frechet_mean(manifold_sphere(), x, ...)`}
\usage{
mean_on_sphere(
  x,
  tau = 0.1,
  tol = 1e-08,
  max.iter = 1000,
  init = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{an (n by q) array of data}

\item{tau}{step size}

\item{tol}{convergence tolerance}

\item{max.iter}{maximum number of iterations}

\item{init}{optional initial value}

\item{verbose}{if TRUE, print the loss at each iteration}
}
\description{
Legacy alias for `frechet_mean(manifold_sphere(), x, ...)`
}
\seealso{
[frechet_mean()]
}
