% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_transport.R
\name{parallel_transport}
\alias{parallel_transport}
\title{Parallel transport of a tangent vector along a geodesic}
\usage{
parallel_transport(mfd, from, to, v, ...)
}
\arguments{
\item{mfd}{a manifold object}

\item{from}{starting point}

\item{to}{end point}

\item{v}{tangent vector at `from`}

\item{...}{manifold-specific extra arguments (e.g. `method` for manifold_bws())}
}
\description{
Parallel transport of a tangent vector along a geodesic
}
