% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{pt_bws}
\alias{pt_bws}
\title{Legacy alias for `parallel_transport(manifold_bws(), Sigma1, Sigma2, V, method)`}
\usage{
pt_bws(Sigma1, Sigma2, V, method = "adams")
}
\arguments{
\item{Sigma1}{starting point}

\item{Sigma2}{end point}

\item{V}{tangent vector at Sigma1}

\item{method}{ODE solver method passed to `deSolve::ode`}
}
\description{
Legacy alias for `parallel_transport(manifold_bws(), Sigma1, Sigma2, V, method)`
}
\seealso{
[parallel_transport()]
}
