% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws.R
\name{pt_bws_core}
\alias{pt_bws_core}
\title{Parallel transport in the Bures-Wasserstein space along the geodesic}
\usage{
pt_bws_core(Sigma1, Sigma2, V, method = "adams")
}
\arguments{
\item{Sigma1}{starting point}

\item{Sigma2}{end point}

\item{V}{tangent vector, identified as a symmetric matrix, at the starting point}
}
\description{
Parallel transport in the Bures-Wasserstein space along the geodesic
}
