% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grassmann.R
\name{pt_gr_core}
\alias{pt_gr_core}
\title{Parallel transport via geodesics on the Grassmannian}
\usage{
pt_gr_core(P, W, V)
}
\arguments{
\item{P}{starting point}

\item{W}{tangent vector associated with the end point (Log_start End)}

\item{V}{tangent vector at P (a matrix)}
}
\description{
Parallel transport via geodesics on the Grassmannian
}
