% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{pt_sphere}
\alias{pt_sphere}
\title{Legacy alias for `parallel_transport(manifold_sphere(), x, y, V)`}
\usage{
pt_sphere(x, y, V)
}
\arguments{
\item{x}{starting point}

\item{y}{end point}

\item{V}{tangent vector at x (a vector or k by d array)}
}
\description{
Legacy alias for `parallel_transport(manifold_sphere(), x, y, V)`
}
\seealso{
[parallel_transport()]
}
