% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{pt_sphere_core}
\alias{pt_sphere_core}
\title{Parallel transport via geodesics for the sphere}
\usage{
pt_sphere_core(x, y, V)
}
\arguments{
\item{x}{starting point}

\item{y}{end point}

\item{V}{tangent vector at x (a vector or k by d array)}
}
\description{
Parallel transport via geodesics for the sphere
}
