% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws_factor_model.R
\name{rfm_bws}
\alias{rfm_bws}
\title{Estimate the RFM in Bures-Wasserstein manifold}
\usage{
rfm_bws(x, r, h = 6, batch_size = NULL, max.iter = 100, mu_hat = NULL)
}
\arguments{
\item{x}{an $n \times p \times p$ array of data where sample is a $p \times p$ SPD matrix}

\item{r}{number of factors to extract}

\item{h}{number of lags used in estimating the factor model}
}
\value{
\describe{
 \item{A}{Estimated loading matrix}
 \item{f_hat}{Estimated factor process}
 \item{E}{A set of orthonormal basis for the tangent space}
 \item{E_lyapunov}{A set of orthonormal basis passed to the Lyapunov operator}
 \item{mu_hat}{Estimated Frechet mean}
}
}
\description{
This function estimates the Riemannian factor model (RFM) with data being
symmetric positive definite matrices equipped with the Bures-Wasserstein
metric.
}
\examples{
set.seed(1)
n = 30; p = 3
X = array(NA, dim = c(n, p, p))
for (i in 1:n) {
  A = matrix(rnorm(p * p), p, p)
  X[i, , ] = A \%*\% t(A) + diag(p)
}

model = rfm_bws(X, r = 1, h = 3)
model$mu_hat          # estimated Frechet mean
dim(model$f_hat)       # estimated factor process, n by r

# evaluate fraction of variance explained on (here, in-sample) data
Frac_Var_bws(X[1:5, , ], model, Euclidean_mean = model$mu_hat)
}
