% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\name{tangent_from_coords}
\alias{tangent_from_coords}
\title{Reconstruct tangent matrix/matrices from coordinates w.r.t. a basis}
\usage{
tangent_from_coords(basis, coords)
}
\arguments{
\item{basis}{an (m by p by p) array of basis matrices}

\item{coords}{a length-m coefficient vector, or an (n by m) matrix}
}
\value{
a (p by p) matrix, or an (n by p by p) array if `coords` is a matrix
}
\description{
Unifies the (mathematically identical) coord2tan()/log_to_tangent() from
the old Grassmannian/BWS utilities.
}
\examples{
Sigma = diag(3) + 0.1
basis = tan_basis_bws(Sigma)$E
z = c(1, 0, 0, 0, 0, 0)          # coordinates of the first basis element
tangent_from_coords(basis, z)    # reconstructs basis[1, , ]
}
