% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\name{coords_from_basis}
\alias{coords_from_basis}
\title{Coordinates of tangent matrix/matrices w.r.t. a basis}
\usage{
coords_from_basis(basis, x)
}
\arguments{
\item{basis}{an (m by p by p) array of basis matrices}

\item{x}{a (p by p) matrix or an (n by p by p) array}
}
\description{
Unifies the (mathematically identical) coord_gr()/tangent_in_E() from the
old Grassmannian/BWS utilities: for each basis element E_i, computes
0.5 * trace(E_i %*% x).
}
