% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws_factor_model.R
\name{Frac_Var_bws}
\alias{Frac_Var_bws}
\title{Computes evaluation metric for RFM on BWS}
\usage{
Frac_Var_bws(
  x_test,
  RFM_model,
  evaluation_type = "BWS",
  fraction = TRUE,
  Euclidean_mean,
  return_predictions = FALSE
)
}
\arguments{
\item{x_test}{raw test data}

\item{RFM_model}{output from rfm_bws}

\item{evaluation_type}{to compute BWS distance or Euclidean (Frobenius distance)}

\item{fraction}{whether to return fraction of variance unexplained or squared prediction errors}
}
\value{
a length-r vector, the fraction of variance unexplained (or squared prediction error) using 1..r factors
}
\description{
Computes evaluation metric for RFM on BWS
}
\examples{
set.seed(1)
n = 30; p = 3
X = array(NA, dim = c(n, p, p))
for (i in 1:n) {
  A = matrix(rnorm(p * p), p, p)
  X[i, , ] = A \%*\% t(A) + diag(p)
}
model = rfm_bws(X, r = 1, h = 3)
Frac_Var_bws(X[1:5, , ], model, Euclidean_mean = model$mu_hat)
}
