% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\name{Log_map}
\alias{Log_map}
\title{Logarithm map on a manifold}
\usage{
Log_map(mfd, x, mu)
}
\arguments{
\item{mfd}{a manifold object}

\item{x}{a single point or a batch of points}

\item{mu}{base point}
}
\description{
Logarithm map on a manifold
}
