% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\name{manifold_grassmann}
\alias{manifold_grassmann}
\title{Grassmannian Gr(d, p)}
\usage{
manifold_grassmann(p = NULL)
}
\arguments{
\item{p}{rank of the projectors; if NULL it is inferred per-call from the data}
}
\description{
Grassmannian Gr(d, p)
}
