% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_transport.R
\name{parallel_transport}
\alias{parallel_transport}
\title{Parallel transport of a tangent vector along a geodesic}
\usage{
parallel_transport(mfd, from, to, v, ...)
}
\arguments{
\item{mfd}{a manifold object}

\item{from}{starting point}

\item{to}{end point}

\item{v}{tangent vector at `from`}

\item{...}{manifold-specific extra arguments (e.g. `method` for manifold_bws())}
}
\value{
the transported tangent vector, at `to`
}
\description{
Parallel transport of a tangent vector along a geodesic
}
\examples{
mfd = manifold_sphere()
x = c(1, 0, 0)
y = c(0, 1, 0)
v = c(0, 0, 1)   # tangent to the sphere at x
parallel_transport(mfd, x, y, v)
}
