% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws.R
\name{tan_basis_bws}
\alias{tan_basis_bws}
\title{Computes a canonical basis for the tangent space at a point in BWS}
\usage{
tan_basis_bws(Sigma)
}
\arguments{
\item{Sigma}{Base point}
}
\value{
An array of orthonormal basis (E), and an array of E passed through the Lyapunov operator
}
\description{
Computes a canonical basis for the tangent space at a point in BWS
}
